Allow the validation tools to detect Lexus versions
