+++ .gitignore
@@ -30,6 +30,7 @@ build/Release
 # Dependency directory
 # https://docs.npmjs.com/misc/faq#should-i-check-my-node-modules-folder-into-git
 node_modules
+package-lock.json
 
 ## Directory-based project format:
 .idea
