--- schema/definitions/filters.json
@@ -1,176 +1,0 @@
-{
-  "title": "Document Filtering",
-  "description": "Filter operators to limit the matching result set before method invocation",
-  "type": "object",
-  "additionalProperties": false,
-  "properties": {
-    "$and": {
-      "title": "AND-style query clauses",
-      "description": "Dictates that documents much match all child clauses",
-      "type": "array",
-      "items": [
-        {
-        }
-      ]
-    },
-    "$or": {
-      "title": "OR-style query clauses",
-      "description": "Dictates that documents much match any child clauses",
-      "type": "array",
-      "items": [
-        {
-        }
-      ]
-    },
-    "$not": {
-      "title": "NOT-style query clauses",
-      "description": "Dictates that documents much match no child clauses",
-      "type": "array",
-      "items": [
-        {
-        }
-      ]
-    },
-    "$exists": {
-      "title": "Existence Filters",
-      "description": "Filters to verify whether fields exist in documents",
-      "type": "object",
-      "additionalProperties": {
-        "type": "boolean",
-        "description": "Boolean value to specify whether the field should exist"
-      }
-    },
-    "$geo": {
-      "title": "Geographic Polygon Filters",
-      "description": "Filters based around geographic locations",
-      "type": "object",
-      "additionalProperties": {
-        "type": "array",
-        "description": "Coordinates to apply as bounds of a polygon, in the form of longitude/latitude pairs",
-        "minItems": 3,
-        "items": {
-          "type": "object",
-          "additionalProperties": false,
-          "properties": {
-            "lat": {
-              "type": "number",
-              "minimum": -90,
-              "maximum": 90
-            },
-            "lon": {
-              "type": "number",
-              "minimum": -180,
-              "maximum": 180
-            }
-          }
-        }
-      }
-    },
-    "$match": {
-      "title": "Matching Filters",
-      "description": "Abitrary filters to compare values for equality",
-      "type": "object",
-      "additionalProperties": {
-        "oneOf": [
-          {
-            "type": "array",
-            "description": "Matching clauses in a standard OR-style"
-          },
-          {
-            "type": "string",
-            "description": "Arbitrary value the field must match"
-          },
-          {
-            "type": "number",
-            "description": "Arbitrary value the field must match"
-          }
-        ]
-      }
-    },
-    "$prefix": {
-      "title": "Prefix Filters",
-      "description": "Arbitrary strings to locate at the start of a field value",
-      "type": "object",
-      "additionalProperties": {
-        "oneOf": [
-          {
-            "type": "array",
-            "description": "Prefix clauses in a standard OR-style",
-            "items": {
-              "type": "string"
-            }
-          },
-          {
-            "type": "string",
-            "description": "Single prefix the value must begin with"
-          }
-        ]
-      }
-    },
-    "$range": {
-      "title": "Range Filtering",
-      "description": "Filters on scalable values with plausible bounds",
-      "type": "object",
-      "additionalProperties": {
-        "type": "object",
-        "description": "Range bounds to apply to the specified field",
-        "additionalProperties": {
-          "anyOf": [
-            {
-              "anyOf": [
-                {
-                  "lt": {
-                    "type": "number",
-                    "description": "Exclusive upper bound designation"
-                  }
-                },
-                {
-                  "lte": {
-                    "type": "number",
-                    "description": "Inclusive upper bound designation"
-                  }
-                }
-              ]
-            },
-            {
-              "anyOf": [
-                {
-                  "gt": {
-                    "type": "number",
-                    "description": "Exclusive lower bound designation"
-                  }
-                },
-                {
-                  "gte": {
-                    "type": "number",
-                    "description": "Inclusive lower bound designation"
-                  }
-                }
-              ]
-            }
-          ]
-        }
-      }
-    },
-    "$suffix": {
-      "title": "Suffix Filters",
-      "description": "Arbitrary strings to locate at the end of a field value",
-      "type": "object",
-      "additionaProperties": {
-        "oneOf": [
-          {
-            "type": "array",
-            "description": "Suffix clauses in a standard OR-style",
-            "items": {
-              "type": "string"
-            }
-          },
-          {
-            "type": "string",
-            "description": "Single suffix the value must end with"
-          }
-        ]
-      }
-    }
-  }
-}
