--- schema/definitions/groups.yml
@@ -1,36 +1,0 @@
-description: A grouping object, dictating the type and field to group on
-type: array
-items:
-  oneOf:
-    - type: string
-      description: A field name representing terms to group by
-    - type: object
-      description: A grouping object, dictating the type and field to group on
-      required:
-      - field
-      - type
-      additionalProperties: false
-      properties:
-        type:
-          title: Grouping Type
-          description: The data type being used by the grouping operation
-          oneOf:
-          - enum:
-            - date
-            - number
-            - string
-          - type: string
-            pattern: ^custom:\w+$
-        field:
-          title: Grouping Field
-          description: The document path to look at when creating groups
-          type: string
-        params:
-          title: Grouping Parameters
-          description:
-            Arbitrary parameters to pass to the grouping operation. Behaviour
-            of these options is specific to the backend being utilised.
-          oneOf:
-          - "$ref": groups/date.json
-          - "$ref": groups/number.json
-          - "$ref": groups/string.json
