--- schema/definitions/groups/number.json
@@ -1,16 +1,0 @@
-{
-  "type": "object",
-  "required": [
-    "interval"
-  ],
-  "properties": {
-    "interval": {
-      "title": "Numeric Interval",
-      "description": "Numeric range to use when grouping values into buckets"
-    },
-    "default": {
-      "title": "Default Value",
-      "description": "A default value to apply to documents in groupings"
-    }
-  }
-}
