--- schema/definitions/invoke.json
@@ -1,47 +1,0 @@
-{
-  "title": "Query Operations",
-  "description": "The actions to take against the dataset",
-  "type": "object",
-  "required": [
-    "method"
-  ],
-  "allOf": [
-    {
-      "properties": {
-        "field": {
-          "title": "Target field designation",
-          "description": "The document path being operated on.",
-          "type": "string"
-        }
-      }
-    },
-    {
-      "oneOf": [
-        {
-          "$ref": "invoke/avg.json"
-        },
-        {
-          "$ref": "invoke/cardinality.json"
-        },
-        {
-          "$ref": "invoke/count.json"
-        },
-        {
-          "$ref": "invoke/find.json"
-        },
-        {
-          "$ref": "invoke/max.json"
-        },
-        {
-          "$ref": "invoke/min.json"
-        },
-        {
-          "$ref": "invoke/percentiles.json"
-        },
-        {
-          "$ref": "invoke/sum.json"
-        }
-      ]
-    }
-  ]
-}
