--- schema/definitions/invoke/avg.json
@@ -1,12 +1,0 @@
-{
-  "required": [ "field" ],
-  "properties": {
-    "method": {
-      "type": "string",
-      "const": "avg"
-    },
-    "params": {
-      "type": "object"
-    }
-  }
-}
