--- schema/definitions/invoke/cardinality.json
@@ -1,12 +1,0 @@
-{
-  "required": [ "field" ],
-  "properties": {
-    "method": {
-      "type": "string",
-      "const": "cardinality"
-    },
-    "params": {
-      "type": "object"
-    }
-  }
-}
