--- schema/definitions/invoke/count.json
@@ -1,12 +1,0 @@
-{
-  "required": [ "field" ],
-  "properties": {
-    "method": {
-      "type": "string",
-      "const": "count"
-    },
-    "params": {
-      "type": "object"
-    }
-  }
-}
