--- schema/definitions/invoke/find.json
@@ -1,43 +1,0 @@
-{
-  "properties": {
-    "method": {
-      "type": "string",
-      "const": "find"
-    },
-    "params": {
-      "type": "object",
-      "properties": {
-        "limit": {
-          "type": "number",
-          "description": "A limit on the number of returned documents"
-        },
-        "offset": {
-          "type": "number",
-          "description": "A number of documents to skip before finding documents to return"
-        },
-        "sort": {
-          "type": "object",
-          "additionalProperties": {
-            "type": "number",
-            "description": "A sort index to apply to the designated field",
-            "enum": [ -1, 1 ]
-          }
-        },
-        "include": {
-          "type": "array",
-          "description": "A list of fields to include in the returned documents",
-          "items": {
-            "type": "string"
-          }
-        },
-        "exclude": {
-          "type": "array",
-          "description": "A list of fields to exclude from the returned documents",
-          "items": {
-            "type": "string"
-          }
-        }
-      }
-    }
-  }
-}
