--- schema/definitions/invoke/max.json
@@ -1,12 +1,0 @@
-{
-  "required": [ "field" ],
-  "properties": {
-    "method": {
-      "type": "string",
-      "const": "max"
-    },
-    "params": {
-      "type": "object"
-    }
-  }
-}
