--- schema/definitions/invoke/min.json
@@ -1,12 +1,0 @@
-{
-  "required": [ "field" ],
-  "properties": {
-    "method": {
-      "type": "string",
-      "const": "min"
-    },
-    "params": {
-      "type": "object"
-    }
-  }
-}
