--- schema/definitions/invoke/percentiles.json
@@ -1,12 +1,0 @@
-{
-  "required": [ "field" ],
-  "properties": {
-    "method": {
-      "type": "string",
-      "const": "percentiles"
-    },
-    "params": {
-      "type": "object"
-    }
-  }
-}
