--- schema/definitions/invoke/sum.json
@@ -1,12 +1,0 @@
-{
-  "required": [ "field" ],
-  "properties": {
-    "method": {
-      "type": "string",
-      "const": "sum"
-    },
-    "params": {
-      "type": "object"
-    }
-  }
-}
