--- schema/schema.json
@@ -1,29 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "description": "Lexus Query Specification",
-  "type": "object",
-  "required": [
-    "version",
-    "invoke"
-  ],
-  "additionalProperties": false,
-  "properties": {
-    "version": {
-      "title": "Specification Version",
-      "description": "The version identifier of the specification model",
-      "type": "string"
-    },
-    "invoke": {
-      "title": "Query operations",
-      "$ref": "definitions/invoke.json"
-    },
-    "filters": {
-      "title": "Document Filtering",
-      "$ref": "definitions/filters.json"
-    },
-    "groups": {
-      "title": "Document Grouping",
-      "$ref": "definitions/groups.yml"
-    }
-  }
-}
