+++ suite/queries/query.avgExists.json
