+++ suite/queries/query.maxGroup.json
