+++ suite/queries/query.minSuffixGroup.json
